requirements.txt
neo4j
python-dotenv
pyyaml
mcp
matplotlib